move create new data to a hook
